# mobile/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterInput"
      responses:
        "200":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /trips/{id}/replies:
    post:
      operationId: createTripReply
      tags: [trips]
      summary: Reply to a trip post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripReplyInput"
      responses:
        "201":
          description: Reply created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripReplyEnvelope"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book seats on a trip (atomically decrements trip seats)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingEnvelope"
        "400":
          description: Invalid input or no seats available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bookings/mine:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: List the authenticated rider's bookings
      responses:
        "200":
          description: Booking list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /support/auth/login:
    post:
      operationId: supportLogin
      tags: [support-auth]
      summary: Agent login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportAgent"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SupportLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    SupportAgent:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, agent]
        createdAt:
          type: string
          format: date-time
...
    AuthUser:
      type: object
      required:
        - id
        - name
        - email
        - phone
        - role
        - rating
        - trips
        - isVerified
        - onboarded
        - isFoundingMember
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          oneOf:
            - type: string
              enum: [driver, rider]
            - type: "null"
        rating:
          type: number
        trips:
          type: integer
        isVerified:
          type: boolean
        onboarded:
          type: boolean
        isFoundingMember:
          type: boolean
          description: True for the first 10,000 users. Grants founding-member badge + crown avatar.
        subscriptionStatus:
          oneOf:
            - type: string
              enum: [active, trialing, past_due, canceled, unpaid, incomplete, incomplete_expired, paused, none]
            - type: "null"
          description: Current Stripe subscription status, or null if no subscription exists.
        trialEndsAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
          description: ISO timestamp when the trial ends (for trialing subscriptions).
...
    AuthSession:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        token:
          type: string
    AuthLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    AuthRegisterInput:
      type: object
      required: [name, email, phone, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
...
    TripReply:
      type: object
      required: [id, userId, userName, text, isDriverReply, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        userName: { type: string }
        text: { type: string }
        isDriverReply: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    TripReplyEnvelope:
      type: object
      required: [reply]
      properties:
        reply: { $ref: "#/components/schemas/TripReply" }
...
    CreateTripReplyInput:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 1000 }
    BookingTripSummary:
      type: object
      required: [id, fromCity, toCity, departureAt, driverName, car]
      properties:
        id: { type: string }
        fromCity: { type: string }
        toCity: { type: string }
        departureAt: { type: string, format: date-time }
        driverName: { type: string }
        car: { type: string }
    Booking:
      type: object
      required:
        - id
        - tripId
        - riderId
        - seats
        - pricePerSeat
        - serviceFee
        - totalAmount
        - paymentMethod
        - status
        - createdAt
        - trip
      properties:
        id: { type: string }
        tripId: { type: string }
        riderId: { type: string }
        seats: { type: integer }
        pricePerSeat: { type: number }
        serviceFee: { type: number }
        totalAmount: { type: number }
        paymentMethod:
          type: string
          enum: [card, apple, venmo]
        status:
          type: string
          enum: [confirmed, cancelled, completed]
        createdAt: { type: string, format: date-time }
        completedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        trip:
          $ref: "#/components/schemas/BookingTripSummary"
    BookingEnvelope:
      type: object
      required: [booking]
      properties:
        booking: { $ref: "#/components/schemas/Booking" }
    BookingList:
      type: object
      required: [bookings]
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
    CreateBookingInput:
      type: object
      required: [tripId, seats, paymentMethod]
      properties:
        tripId: { type: string, minLength: 1 }
        seats: { type: integer, minimum: 1, maximum: 8 }
        paymentMethod:
          type: string
          enum: [card, apple, venmo]